==> code/00_reddit_data.ipynb <==
# Getting data

In this notebook, we'll be using the function defined in `pushshift_functions.py` to grab posts from subreddits related to COVID-19.

## Imports

import os

from pushshift_functions import create_combined_df

## Using the function

# Create list of subreddits to pull from.
subreddits = [
    'CoronaVirusPA',
    'coronavirusSC',
    'COVID19',
    'China_Flu',
    'CoronavirusRecession',
    'CoronavirusAtlanta',
    'nCoV',
    'CoronavirusNJ',
    'CoronavirusNE',
    'CoronavirusAlabama',
    'CoronaNC',
    'CoronaVirusSouthEast',
    'coronavirusflorida',
    'coronavirusVA',
    'CoronaVirusWV',
    'CoronaVirus_2019_nCoV',
    'Trumpvirus',
    'CoronavirusAustralia',
    'CoronavirusFOS',
    'CoronavirusUS',
    'CoronavirusUK',
    'CoronavirusDownunder',
    'CanadaCoronavirus',
    'CoronavirusCA',
    'FloridaCoronavirus',
    'CoronavirusNewYork',
    'CoronaVirusTX',
    'CoronavirusWA',
    'CoronavirusMichigan',
    'CoronavirusColorado',
    'CoronavirusIllinois',
    'Coronaviruslouisiana',
    'CoronavirusMa',
    'CoronavirusGA'
]

# The following grabs every post ever made from each of the given subreddits.
# This will take a LONG time to run.
df = create_combined_df(subreddits)

## Examining the data

df.head()

df.shape

## Exporting the raw data

try:
    os.mkdir('../data')
except FileExistsError:
    pass

df.to_csv(f'../data/reddit_raw.csv', index=False)
