==> code/01_reddit_cleaning.ipynb <==
# Cleaning

## Imports

import os
import pandas as pd
import time

## Loading the data

df = pd.read_csv('../data/reddit_raw.csv', low_memory=False)

df.head()

df.shape

## Clean the raw dataframe

# Overwrite dataframe to be only the columns of potential interest
features = ['subreddit', 
            'subreddit_subscribers', 
            'title', 
            'selftext', 
            'score', 
            'num_comments', 
            'author', 
            'created_utc']
df = df[features]

# Select only the subreddits related to geographic regions
subreddits = [
    'CoronavirusUS',
    'CoronavirusCA',
    'CoronavirusWA',
    'CoronaVirusTX',
    'CoronavirusNewYork',
    'CoronavirusGA',
    'FloridaCoronavirus',
    'CoronavirusMichigan',
    'CoronavirusColorado',
    'CoronavirusMa',
    'CoronavirusAlabama',
    'CoronavirusIllinois',
    'Coronaviruslouisiana',
    'CoronaVirusPA',
    'coronavirusSC',
    'coronavirusVA',
    'CoronaVirusWV',
    'CoronavirusNJ'
]
df = df.loc[df['subreddit'].isin(subreddits), :]

# Convert utc column to regular date format
def get_date_from_utc(utc_time):
    local_time = time.localtime(utc_time)
    return str(local_time.tm_year) + '-' + str(local_time.tm_mon) + '-' + str(local_time.tm_mday)

df['date'] = pd.to_datetime(df['created_utc'].map(get_date_from_utc))

df.dtypes

## Exporting the clean data

try:
    os.mkdir('../data')
except FileExistsError:
    pass

df.to_csv(f'../data/reddit_clean.csv', index=False)
