==> code/02_combining_dataframes.ipynb <==
## Imports

import os
import pandas as pd

## Loading the data

reddit = pd.read_csv('../data/reddit_clean.csv')

us = pd.read_csv('../data/us.csv')
us.set_index('date', inplace=True)

states = pd.read_csv('../data/us-states.csv')
states.set_index('date', inplace=True)

## Amalgamate us data

# Create dataframe of only reddit posts from CoronavirusUS
reddit_us = reddit.loc[reddit['subreddit'] == 'CoronavirusUS', :]

# Add subreddit column to us dataframe
us['subreddit'] = 'CoronavirusUS'

# Add total number of subreddit subscribers by date to us dataframe
us = us.merge(reddit_us.groupby('date')[['subreddit_subscribers']].max(), 
              left_index=True,
              right_index=True)

# Add total number of subreddit posts by date to us dataframe
us = us.merge(reddit_us.groupby('date')[['title']].count(),
              left_index=True,
              right_index=True).rename(columns={'title': 'num_posts'})

# Add total number of subreddit comments by date to us dataframe
us = us.merge(reddit_us.groupby('date')[['num_comments']].sum(), 
              left_index=True,
              right_index=True)

# Add total score of subreddit posts by date to us dataframe
us = us.merge(reddit_us.groupby('date')[['score']].sum(), 
              left_index=True,
              right_index=True).rename(columns={'score': 'tot_score'})

## Amalgamate state data

# Create dict of states and their corresponding subreddits
states_and_subreddits = {
    'Pennsylvania': 'CoronaVirusPA',
    'South Carolina': 'coronavirusSC',
    'New Jersey': 'CoronavirusNJ',
    'Alabama': 'CoronavirusAlabama',
    'Virginia': 'coronavirusVA',
    'West Virginia': 'CoronaVirusWV',
    'California': 'CoronavirusCA',
    'Florida': 'FloridaCoronavirus',
    'New York': 'CoronavirusNewYork',
    'Texas': 'CoronaVirusTX',
    'Washington': 'CoronavirusWA',
    'Michigan': 'CoronavirusMichigan',
    'Colorado': 'CoronavirusColorado',
    'Illinois': 'CoronavirusIllinois',
    'Louisiana': 'Coronaviruslouisiana',
    'Massachusetts': 'CoronavirusMa',
    'Georgia': 'CoronavirusGA',
}

state_dataframes = []

for state in states_and_subreddits.keys():
    # Create dataframe of only reddit posts from the given state
    state_reddit = reddit.loc[reddit['subreddit'] == states_and_subreddits[state], :]

    # Create dataframe of only cases and deaths from the given state.
    state_df = states.loc[states['state'] == state, :]
    
    # Add subreddit column to state dataframe
    state_df['subreddit'] = states_and_subreddits[state]

    # Add total number of subreddit subscribers by date to state dataframe
    state_df = state_df.merge(state_reddit.groupby('date')[['subreddit_subscribers']].max(), 
                              left_index=True,
                              right_index=True)

    # Add total number of subreddit posts by date to state dataframe
    state_df = state_df.merge(state_reddit.groupby('date')[['title']].count(), 
                              left_index=True,
                              right_index=True).rename(columns={'title': 'num_posts'})

    # Add total number of subreddit comments by date to state dataframe
    state_df = state_df.merge(state_reddit.groupby('date')[['num_comments']].sum(), 
                              left_index=True,
                              right_index=True)

    # Add total score of subreddit posts by date to us dataframe
    state_df = state_df.merge(state_reddit.groupby('date')[['score']].sum(), 
                              left_index=True,
                              right_index=True).rename(columns={'score': 'tot_score'})
    
    state_dataframes.append(state_df)

df = pd.concat(state_dataframes)

df[df['state'] == 'West Virginia'].tail()

## Export data

try:
    os.mkdir('../data')
except FileExistsError:
    pass
finally:
    try:
        os.mkdir('../data/states')
    except FileExistsError:
        pass

us.to_csv('../data/us_with_reddit.csv')

for state_df in state_dataframes:
    filename = state_df['state'][0].lower().replace(' ', '_') + '.csv'
    state_df.to_csv(f'../data/states/{filename}')

pd.concat(state_dataframes).to_csv('../data/states_with_reddit.csv')
