==> code/03_new_jersey.ipynb <==
## Imports

import pandas as pd
import time

from pushshift_functions import get_subreddit_posts

## Pulling from /r/newjersey

# Create list to hold all the posts we acquire
posts = []

# Create count variable to keep track of pulls
count = 0
print(f'Pull number: {count + 1}')

# Make initial pull of reddit posts.
# after is set to one year ago, start pulling posts from then.
data = get_subreddit_posts({
        'subreddit': 'newjersey',
        'size': 500,
        'after': 1557769802,
})

# As long as the pull of posts was successful, the following should run.
# Keep pulling posts as long as there are more to grab.
while data:
    posts.append(data)

    # By design, the newest post will always be from the last list in posts.
    data = get_subreddit_posts({
        'subreddit': 'newjersey',
        'size': 500,
        'after': max([post['created_utc'] for post in posts[-1]])
,
    })
    
    count += 1
    print(f'Pull number: {count + 1}')

    time.sleep(5)

dfs = [pd.DataFrame(bundle) for bundle in posts]

df = pd.concat(dfs, ignore_index=True)

## Export data

df.to_csv('../data/new_jersey_subreddit.csv')
