==> code/04_twitter_cleaning.ipynb <==
## Imports

import pandas as pd

from langdetect import detect

## Load the data

df = pd.read_csv('tweets.csv')

## Cleaning

languages = []
error_indices = []
for i, text in enumerate(df['text']):
    try:
        languages.append(detect(text))
    except:
        error_indices.append(i)

df.drop(index=error_indices, inplace=True)
df.reset_index(drop=True, inplace=True)

df['language'] = languages

not_english_indices = df.loc[df['language'] != 'en', :].index

df.drop(index=not_english_indices, inplace=True)
df.reset_index(drop=True, inplace=True)

df.drop(columns='language', inplace=True)

## Export the data

df.to_csv('./english_tweets.csv', index=False)
